requirements/base.txt
Django==5.1.3
python-environ==0.4.54
Pillow==11.0.0
djangorestframework==3.15.2